CosmoCentral.jl

Documentation for CosmoCentral.jl. This is a Julia package to perform cosmological calculations.

Authors

Marco Bonici, Dipartimento di Fisica, Università degli Studi di Genova (UniGe)

License

CosmoCentral is licensed under the MIT "Expat" license; see LICENSE for the full license text.

CosmoCentral.jl Documentation

This a provisional Documentation.

CosmoCentral.ComputeAdimensionalHubbleFactor — Function

ComputeAdimensionalHubbleFactor($z$, params)

This function, given the value of the cosmological parameters, evaluate the Adimensional Hubble Factor for $w_0 w_a$CDM cosmologies.

...

Arguments

z::Float64 the redshift value at which evaluate the Adimensional Hubble Factor

params::w0waCDMCosmology a collection of cosmological parameters, whose expression is given by

\[E(z)=\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Notes

This expression is valid only for the CPL parameterization

CosmoCentral.ComputeComovingDistance — Function

ComputeComovingDistance($z$, params)

This function, given the value of the cosmological parameters, evaluate the Comoving Distance. It is evaluated as

\[r(z)=\frac{c}{H_0}\int_0^z \frac{dx}{E(x)}\]