Cosmological Structures

In order to organize the code, some custom types are used. This section is under heavy developement and a lot of things may change.

CosmoCentral.CosmologicalGrid — Type

CosmologicalGrid(ZArray::Vector{Float64} = Array(LinRange(0.001, 2.5, 300)),
KArray::Vector{Float64} = LogSpaced(1e-5, 50., 1000))

This struct contains the value of the Cosmological Grid, both in $k$ and $z$.

CosmoCentral.BackgroundQuantities — Type

BackgroundQuantities(HZArray::Vector{Float64}, χZArray::Vector{Float64}),
DZArray::Vector{Float64}

This struct contains the arrays with the values of the Hubble parameter $H(z)$ and the comoving distance $\chi(z)$.