Cosmology

In this section we show the structures which specify the cosmology and the value of the parameters. In this moment we support only the $w_0 w_a$CDM cosmology, although we plan to include support to other cosmologies.

CosmoCentral.w0waCDMCosmology — Type

w0waCDMCosmology(w0::Float64 = -1, wa::Float64 = 0, ΩM::Float64 = 0.32,
ΩB::Float64  = 0.05, ΩDE::Float64 = 0.68, Ωk::Float64  = 0.,
Ωr::Float64  = 0., ns::Float64  = 0.96, Mν::Float64  = 0.06,
σ8::Float64  = 0.816, H0::Float64  = 67.)

This struct contains the value of the cosmological parameters for $w_0 w_a$CDM cosmologies:

$w_0$ and $w_a$, the parameters in the CPL parameterization

$\Omega_M$, $\Omega_B$, $\Omega_{DE}$, $\Omega_R$, and $\Omega_k$ the density parameters for respectively matter, baryons, Dark Energy, radiation, curvature

$n_s$, the scalar spectral index

$M_\nu$, the sum of the neutrino mass eigenstates in eV

$\sigma_8$, the amplitude of the scalar fluctuations

$H_0$, the value of the Hubble paramater

In the remainder of this documentation, we will show how to use CosmoCentral.jl and we will plot several quantities. The assumed values for the cosmological parameters are:

$\Omega_M$ = 0.32

$\Omega_B$ = 0.05

$\Omega_k$ = 0

$n_s$ = 0.96

$\sigma_8$ = 0.816

$h$ = 0.67

$M_\nu$ = 0.06 eV

$w_0$ = -1

$w_a$ = 0

Cosmological Grid

Another import quantity is the cosmological grid, the structure which specifies the grids used to compute the quantities in our calculations.

CosmoCentral.CosmologicalGrid — Type

CosmologicalGrid{T} <: AbstractCosmologicalGrid{T}
ZArray::AbstractArray{T} = LinRange(0.001, 2.5, 300)
KArray::AbstractArray{T} = LogSpaced(1e-5, 50., 1000)
ℓBinCenters::AbstractArray{T} = LinRange(10., 3000., 2991)
ℓBinWidths::AbstractArray{T} = LinRange(10., 3000., 2991)
KLimberArray::AbstractArray{T,2} = zeros(length(ℓBinCenters),
length(ZArray))
KBeyondLimberArray::AbstractArray{T,2} = zeros(100, 1000)

This struct contains several grids:

ZArray and KArray, the grids for the PowerSpectrum calculations

ℓBinCenters, the $\ell$ grid for the $C_\ell$'s calculations

ℓBinWidths, used when computing the covariance matrix

KLimberArray, the k grid to evaluate the $C_\ell$'s in the Limber approximation

KBeyondLimberArray, the k grid to evaluate the $C_\ell$'s without the Limber approximation