Covariance Matrix

In order to evaluate the Fisher Matrix CosmoCentral.Fisherαβ, a Covariance Matrix is required. Here we show the two Gaussian Covariance Matrix we have implemented, in the Field and Estimator Approach.

Field approach

The Covariance Matrix in the Field approach is evaluated by the following method

CosmoCentral.aₗₘCovariance — Type

aₗₘCovariance()

This struct contains the Covariance in the field perspective, i.e. when the observables are the $a_{\ell m}$'s.

CosmoCentral.InstantiateEvaluateCovariance — Method

InstantiateEvaluateCovariance(cℓ::AbstractCℓ, ConvDens::AbstractConvolvedDensity,
cosmogrid::CosmologicalGrid, ProbeA::String, ProbeB::String)

This function evaluates and returns the aₗₘCovariance, according to the following formula:

\[\Sigma_{i j}^{\mathrm{AB}}(\ell)=\sqrt{\frac{2}{(2 \ell+1) \Delta \ell f_{\mathrm{sky}}}}\left(C_{i j}^{\mathrm{AB}}(\ell)+N_{i j}^{\mathrm{AB}}(\ell) \right),\]

where $\mathrm{A}$ and $\mathrm{B}$ are the probes, $\mathrm{i}$ and $\mathrm{j}$ are the tomographic bins, $\ell$ and $\Delta\ell$ are respectively the central value and width of the considered multipole bin, $f_{\mathrm{sky}}$ is the sky fraction covered by the field considered, $N_{i j}^{\mathrm{AB}}$ is the noise matrix.

Estimator approach

The Fisher Matrix in the Estimator approach is evaluated by the following method

CosmoCentral.CℓCovariance — Type

CℓCovariance()

This struct contains the Covariance in the estimator perspective, i.e. when the observables are the $C_{\ell}$'s.

CosmoCentral.InstantiateEvaluateCovariance — Method

InstantiateEvaluateCovariance(Covaₗₘ::aₗₘCovariance)

This function evaluates and returns the CℓCovariance, using the following formula:

\[\boldsymbol{\Xi}(\ell)= \left(\boldsymbol{D}_{n}^{T}\left( \left(\boldsymbol{\Sigma}(\ell)\right)^{-1} \otimes \left(\boldsymbol{\Sigma}(\ell)\right)^{-1}\right) \boldsymbol{D}_{n}\right)^{-1}\]

where $\boldsymbol{D}_{n}$ is the DuplicationMatrix.