The CosmoCentral Module

Module Index

CosmoCentral.w0waCDMParameters

CosmoCentral.ComputeAdimensionalHubbleFactor

CosmoCentral.ComputeComovingDistance

CosmoCentral.ComputeDensityFunction

CosmoCentral.ComputeHubbleFactor

CosmoCentral.ComputeInstrumentResponse

CosmoCentral.NormalizeAnalitycalDensityStruct

CosmoCentral.NormalizeConvolvedDensityStruct

Detailed API

CosmoCentral.w0waCDMParameters — Type

w0waCDMParameters

This struct contains the value of the cosmological parameters for $w_0 w_a$CDM cosmologies:

$w_0$ and $w_a$, the parameters in the CPL parameterization

$\Omega_M$, $\Omega_B$, $\Omega_{DE}$, $\Omega_R$, and $\Omega_k$ the density parameters for matter, baryons, Dark Energy, radiation, curvature

$n_s$, the scalar spectral index

$M_\nu$, the sum of the neutrino mass eigenstates in eV

$\sigma_8$, the amplitude of the scalar fluctuations

$H_0$, the value of the Hubble paramater

CosmoCentral.ComputeAdimensionalHubbleFactor — Method

ComputeAdimensionalHubbleFactor($z$, params)

This function, given the value of the cosmological parameters, evaluate the Adimensional Hubble Factor for $w_0 w_a$CDM cosmologies.

...

Arguments

z::Float64 the redshift value at which evaluate the Adimensional Hubble Factor

params::w0waCDMCosmology a collection of cosmological parameters, whose expression is given by

\[E(z)=\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Notes

This expression is valid only for the CPL parameterization

CosmoCentral.ComputeComovingDistance — Method

ComputeComovingDistance($z$, params)

This function, given the value of the cosmological parameters, evaluate the Comoving Distance. It is evaluated as

\[r(z)=\frac{c}{H_0}\int_0^z \frac{dx}{E(x)}\]

CosmoCentral.ComputeDensityFunction — Method

ComputeDensityFunction($z$, params)

This function returns the source density for a given redshift $z$

CosmoCentral.ComputeHubbleFactor — Method

ComputeHubbleFactor($z$, params)

...

Arguments

z::Float64 the redshift value at which evaluate the Adimensional Hubble Factor

params::w0waCDMCosmology a collection of cosmological parameters

This function, given the value of the cosmological parameters, evaluate the Hubble Factor for $w_0 w_a$CDM cosmologies, whose expression is given by

\[H(z)=H_0\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Notes

This expression is valid only for the CPL parameterization

CosmoCentral.ComputeInstrumentResponse — Method

ComputeInstrumentResponse($z$, params)

This function returns the instrument response.

CosmoCentral.NormalizeAnalitycalDensityStruct — Method

ComputeDensityFunction(params)

This function modifies the normalization constant in the AnalitycalDensityStruct in order to have the same value of the surface density once integrated.

CosmoCentral.NormalizeConvolvedDensityStruct — Method

ComputeDensityFunction(params)

This function modifies the normalization constant in the AnalitycalDensityStruct in order to have the same value of the surface density once integrated.