Angular Coefficients

Central quantities are the Angular Coefficients $C_\ell$. Actually we implement only the Limber approximation to evaluate the $C_\ell$, according to:

\[C_{i j}^{AB}(\ell)=\frac{c}{H_0} \int_{z_{\min }}^{z_{\max }} \mathrm{d} z \frac{W_{i}^{A}(z) W_{j}^{B}(z)}{E(z) r^{2}(z)} P_{\delta \delta}\left(\frac{\ell+1 / 2}{r(z)}, z\right)\]

CosmoCentral.Cℓ — Type

Cℓ(CℓArray::AbstractArray{Float64, 3})

This struct contains the array with the Angular Coefficients.

CosmoCentral.∂Cℓ — Type

∂Cℓ(∂CℓArray::AbstractArray{Float64, 3})

This struct contains the array with the derivatives of the Angular Coefficients.

CosmoCentral.ComputeCℓ! — Method

ComputeCℓ!(Cℓ::AbstractCℓ, WeightFunctionA::AbstractWeightFunction,
WeightFunctionB::AbstractWeightFunction, BackgroundQuantities::BackgroundQuantities,
::AbstractCosmology, CosmologicalGrid::AbstractCosmologicalGrid,
PowerSpectrum::AbstractPowerSpectrum, ::NumericalIntegrationSimpson)

This function evaluates the Angular Coefficients for all tomographic bins and multipole values. In order to evaluate the numerical integrals, it is used the Simpson numerical method from NumericalIntegration.jl .

CosmoCentral.ComputeCℓ! — Method

ComputeCℓ!(Cℓ::AbstractCℓ, WeightFunctionA::AbstractWeightFunction,
WeightFunctionB::AbstractWeightFunction,
BackgroundQuantities::AbstractBackgroundQuantities,
::AbstractCosmology, CosmologicalGrid::AbstractCosmologicalGrid,
PowerSpectrum::AbstractPowerSpectrum, ::CustomSimpson)

This function evaluates the Angular Coefficients for all tomographic bins and multipole values. In order to evaluate the numerical integrals, it has been implemented the Simpson rule. The computation is accelerated by LoopVectorization.jl .