Angular Coefficients

Central quantities are the Angular Coefficients $C_\ell$. Actually we implement only the Limber approximation to evaluate the $C_\ell$, according to:

\[C_{i j}^{AB}(\ell)=\frac{c}{H_0} \int_{z_{\min }}^{z_{\max }} \mathrm{d} z \frac{W_{i}^{A}(z) W_{j}^{B}(z)}{E(z) r^{2}(z)} P_{\delta \delta}\left(\frac{\ell+1 / 2}{r(z)}, z\right)\]

CosmoCentral.AngularCoefficientsStruct — Type

AngularCoefficientsStruct()

This struct contains the array with the Angular Coefficients.

CosmoCentral.ComputeAngularCoefficients — Function

ComputeAngularCoefficients(AngularCoefficients::AngularCoefficients,
WeightFunctionA::GCWeightFunction, WeightFunctionB::GCWeightFunction,
BackgroundQuantities::BackgroundQuantities,
w0waCDMCosmology::w0waCDMCosmology, CosmologicalGrid::CosmologicalGrid,
PowerSpectrum::PowerSpectrum)

This function evaluates the Angular Coefficients for all tomographic bins and multipole values.