Background Universe

Some of the most basics quantities in cosmology are the Hubble factor $H(z)$ and the comoving distance $r(z)$. In this section are presented the functions which evaluates them.

Hubble factor

CosmoCentral.ComputeAdimensionalHubbleFactor — Function

ComputeAdimensionalHubbleFactor(z::Float64, params::w0waCDMCosmology)

This function, given the value of the cosmological parameters, evaluate the Adimensional Hubble Factor for $w_0 w_a$CDM cosmologies. The analitycal expression is given by:

\[E(z)=\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+ \Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Warning

This expression is valid only for the CPL parameterization of the Dark Energy Equation of State.

CosmoCentral.ComputeHubbleFactor — Function

ComputeHubbleFactor(z::Float64, params::w0waCDMCosmology)

This function, given the value of the cosmological parameters, evaluate the Hubble Factor for $w_0 w_a$CDM cosmologies, whose expression is given by

\[H(z)=H_0\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+ \Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Distances

CosmoCentral.ComputeComovingDistance — Function

ComputeComovingDistance(z::Float64, params::w0waCDMCosmology)

This function, given the value of the cosmological parameters, evaluate the Comoving Distance. It is evaluated as:

\[r(z)=\frac{c}{H_0}\int_0^z \frac{dz'}{E(z')}\]

Utils

CosmoCentral.ComputeBackgroundQuantitiesOverGrid — Function

ComputeBackgroundQuantitiesOverGrid(CosmologicalGrid::CosmologicalGrid,
    BackgroundQuantities::BackgroundQuantities,
    w0waCDMCosmology::w0waCDMCosmology)

This function evaluate the Hubble factor and the comoving distance over the CosmologicalGridStruct.