Bias

One of the most important ingredients of the Galaxy Clustering Weight function is the Galaxy Bias, that is, the statistical relation between the distribution of galaxies and matter

\[\delta_{g}(\boldsymbol{x})=b \delta(\boldsymbol{x})\]

In this page are listed the Bias Model we have implemented.

Piecewise Bias

The Piecewise Bias model is taken from the official Euclid forecast

\[b(z)= \sqrt{1+\bar{z}}\]

where $\bar{z}$ is the redshift value in the center of the tomographic bin where the redshift $z$ lies. The PiecewiseBias is plotted here.

using Plots
using CosmoCentral
CosmologicalGrid = CosmoCentral.CosmologicalGrid(ZArray = LinRange(0.001, 2.5, 500))
ConvolvedDensity = CosmoCentral.ConvolvedDensity(DensityGridArray =
        ones(10, length(CosmologicalGrid.ZArray)))
GCWeightFunction = CosmoCentral.GCWeightFunction(WeightFunctionArray=
        zeros(length(ConvolvedDensity.DensityGridArray[1,:]),
        length(CosmologicalGrid.ZArray)))
CosmoCentral.ComputeBiasGrid!(CosmologicalGrid, GCWeightFunction, ConvolvedDensity)
x = CosmologicalGrid.ZArray; y = GCWeightFunction.BiasArray[1, :];
plot(x, y, label = "Bias", xlabel="z")

Euclid Flagsip Bias

This model, obtained from the Flagship Euclid simulation is taken from Tutusaus et al 2021

\[b(z)=A+\frac{B}{1+\exp [-(z-D) C]}\]

using Plots
using CosmoCentral
CosmologicalGrid = CosmoCentral.CosmologicalGrid(ZArray = LinRange(0.001, 2.5, 500))
ConvolvedDensity = CosmoCentral.ConvolvedDensity(DensityGridArray =
        ones(10, length(CosmologicalGrid.ZArray)))
GCWeightFunction = CosmoCentral.GCWeightFunction(WeightFunctionArray=
        zeros(length(ConvolvedDensity.DensityGridArray[1,:]),
        length(CosmologicalGrid.ZArray)), BiasKind = CosmoCentral.EuclidBias())
CosmoCentral.ComputeBiasGrid!(CosmologicalGrid, GCWeightFunction, ConvolvedDensity)
x = CosmologicalGrid.ZArray; y = GCWeightFunction.BiasArray[1, :];
plot(x, y, label = "Bias", xlabel="z")

CosmoCentral.ComputeBias — Function

ComputeBias(z::Float64, ::PiecewiseBias,
ConvolvedDensity::AbstractConvolvedDensity)

This function evaluate the piecewise bias, given by $\sqrt{1+z}$, for a given redshift.

CosmoCentral.ComputeBiasGrid! — Function

ComputeBiasGrid!(CosmologicalGrid::CosmologicalGrid,
gcWeightFunction::GCWeightFunction, Bias::AbstractBias,
ConvolvedDensity::AbstractConvolvedDensity)

This function evaluate the bias, over the cosmological redshift grid.