BoltzmannSolver

One of the core components of CosmoCentral are the Boltzmann solvers, codes developed by the community that solve the coupled Einstein-Boltzmann equations and used to evaluate the Matter Power Spectrum. Currently, the Boltzmann codes implemented in CosmoCentral are:

CLASS

We plan to include other Boltzmann solvers in the future.

Power Spectrum

CosmoCentral.PowerSpectrum — Type

PowerSpectrum()

This struct contains the array with the Linear and Nonlinear Power Spectrum evaluated on the $k-z$ grid and the interpolated Nonlinear Power Spectrum on Limber $k-z$ grid.

CosmoCentral.ComputeLimberArray! — Function

ComputeLimberArray!(CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities)

This function compute the Limber grid. In the Limber approximation, $k$ and $z$ are related by the following relation:

\[k_\ell(z)=\frac{\ell+1/2}{r(z)}.\]

CosmoCentral.InterpolatePowerSpectrumLimberGrid! — Function

InterpolatePowerSpectrumLimberGrid!(CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities, PowerSpectrum::PowerSpectrum,
::InterpolationMethod)

This function interpolates the Power Spectrum on the $k-z$ grid and evaluates it on the Limber grid.

CosmoCentral.InterpolationMethod — Type

InterpolationMethod

This type is used to specify the method to interpolate the Power Spectrum; actually are included:

Dierckx.jl, the Julia wrapper for the Fortran library Dierckx

GriddedLinear, from Interpolations.jl

BSpliceCubic (recommended, for its speed and accuracy), from Interpolations.jl

classy

classy is the Python wrapper for CLASS.

CosmoCentral.classyParams — Type

classyParams(classyParamsDict::Dict)

This struct contains the dictionary with the classy parameters. For a detalied explanation of the parameters, please refer to the CLASS website

CosmoCentral.Initializeclassy — Function

Initializeclassy(w0waCDMCosmology::w0waCDMCosmology)

This function, given a w0waCDMCosmology, returns the classyParams correctly initialized.

CosmoCentral.EvaluatePowerSpectrum! — Function

EvaluatePowerSpectrum!(classyParams:: classyParams, CosmologicalGrid::CosmologicalGrid,
PowerSpectrum::PowerSpectrum)

This function runs classy to evaluate the Matter Power Spectrum over the $k-z$ grid specified in CosmologicalGrid

For instance, here we show the linear and nonlinear $P_{\delta\delta}(k,z)$, evaluated by CLASS, for a the reference cosmology.

p = plot(CosmologicalGrid.KArray, PowerSpectrum.PowerSpectrumLinArray[:,1],
ylabel = L"P(k)\,\left[\mathrm{Mpc}^{3}\right]", xlabel = L"k\,\left[\mathrm{Mpc}^{-1}\right]",
xaxis=:log, yaxis=:log, label = L"\mathrm{Linear}", xlims = (1e-5,10), ylims = (1,1e5))
plot!(p, CosmologicalGrid.KArray, PowerSpectrum.PowerSpectrumNonlinArray[:,1],
xaxis=:log, yaxis=:log, label = L"\mathrm{Nonlinear}")

When performing forecasts, $P_{\delta \delta}(k,z)$ can be valuated once and stored (we provide a precomputed set of spectra here), so their computational impact is reduced. However, when evaluating $C_\ell$'s, there is a not negligible impact of $P_{ \delta\delta}(k,z)$ interpolation and evaluation on the Limber $k-$grid.

@benchmark CosmoCentral.InterpolatePowerSpectrumLimberGrid!(CosmologicalGrid,
BackgroundQuantities, PowerSpectrum, CosmoCentral.BSplineCubic())

BenchmarkTools.Trial: 117 samples with 1 evaluation.
 Range (min … max):  40.305 ms … 52.382 ms  ┊ GC (min … max): 0.00% … 9.23%
 Time  (median):     41.724 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   43.018 ms ±  2.771 ms  ┊ GC (mean ± σ):  1.60% ± 3.60%

     ▃▅▅█▆▅▃                         ▃   ▁                     
  ▇▅▁███████▇▅▅▁▁▁▁▁▁▁▁▁▁▁▅▅▁▇▁▁▁▁▁▅▅█▅▇▅█▁▅▁▅▁▁▇▁▁▅▁▁▁▁▁▁▁▁▅ ▅
  40.3 ms      Histogram: log(frequency) by time      51.8 ms <

 Memory estimate: 16.45 MiB, allocs estimate: 1148.