BoltzmannSolver

One of the core components of CosmoCentral are the Boltzmann solvers, codes developed by the community that solve the coupled Einstein-Boltzmann equations and used to evaluate the Matter Power Spectrum. Currently, the Boltzmann codes implemented in CosmoCentral are:

CLASS

We plan to include other Boltzmann solvers in the future.

Power Spectrum

CosmoCentral.PowerSpectrumStruct — Type

PowerSpectrumStruct()

This struct contains the array with the Linear and Nonlinear Power Spectrum evaluated on the $k-z$ grid and the interpolated Nonlinear Power Spectrum on Limber $k-z$ grid.

CosmoCentral.ComputeLimberArray — Function

ComputeLimberArray(CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities)

This function compute the Limber grid. In the Limber approximation, $k$ and $z$ are related by the following relation:

\[k_\ell(z)=\frac{\ell+1/2}{r(z)}.\]

CosmoCentral.InterpolateAndEvaluatePowerSpectrum — Function

InterpolateAndEvaluatePowerSpectrum(CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities, PowerSpectrum::PowerSpectrum)

This function interpolates the Power Spectrum on the $k-z$ grid and evaluates it on the Limber grid.

classy

classy is the Python wrapper for CLASS.

CosmoCentral.classyParamsStruct — Type

classyParamsStruct(classyParamsDict::Dict)

This struct contains the dictionary with the classy parameters. For a detalied explanation of the parameters, please refer to the CLASS website

CosmoCentral.Initializeclassy — Function

Initializeclassy(w0waCDMCosmology::w0waCDMCosmology)

This function, given a w0waCDMCosmologyStruct, returns the classyParamsStruct correctly initialized.

CosmoCentral.EvaluatePowerSpectrum — Function

EvaluatePowerSpectrum(classyParams:: classyParams,
CosmologicalGrid::CosmologicalGrid, PowerSpectrum::PowerSpectrum)

This function runs classy to evaluate the Matter Power Spectrum over the $k-z$ grid specified in CosmologicalGridStruct