Cosmological Structure

In order to organize the code, some custom types are used. This section is under heavy developement and a lot of things may change.

CosmoCentral.w0waCDMCosmologyStruct — Type

w0waCDMCosmologyStruct(w0::Float64 = -1, wa::Float64 = 0, ΩM::Float64 = 0.32,
ΩB::Float64  = 0.05, ΩDE::Float64 = 0.68, Ωk::Float64  = 0.,
Ωr::Float64  = 0., ns::Float64  = 0.96, Mν::Float64  = 0.06,
σ8::Float64  = 0.816, H0::Float64  = 67.)

This struct contains the value of the cosmological parameters for $w_0 w_a$CDM cosmologies:

$w_0$ and $w_a$, the parameters in the CPL parameterization

$\Omega_M$, $\Omega_B$, $\Omega_{DE}$, $\Omega_R$, and $\Omega_k$ the density parameters for matter, baryons, Dark Energy, radiation, curvature

$n_s$, the scalar spectral index

$M_\nu$, the sum of the neutrino mass eigenstates in eV

$\sigma_8$, the amplitude of the scalar fluctuations

$H_0$, the value of the Hubble paramater

CosmoCentral.CosmologicalGridStruct — Type

CosmologicalGridStruct(ZArray::Vector{Float64} = Array(LinRange(0.001, 2.5, 300)),
KArray::Vector{Float64} = LogSpaced(1e-5, 50., 1000))

This struct contains the value of the Cosmological Grid, both in $k$ and $z$.

CosmoCentral.BackgroundQuantitiesStruct — Type

BackgroundQuantitiesStruct(HZArray::Vector{Float64},
rZArray::Vector{Float64})

This struct contains the arrays with the values of the Hubble parameter $H(z)$ and the comoving distance $r(z)$.