Weight Function

Galaxy Clustering

The expression of the Galaxy Clustering Weight Function is given by:

\[W_{i}^{\mathrm{G}}(z)=b_{i}(z) n_{i}(z) \frac{H(z)}{c}\]

CosmoCentral.GCWeightFunction — Type

GCWeightFunction()

This struct contains the array with the Galaxy Bias and Galaxy Clustering Weight Function values for all tomographic bins and redshift values in the CosmologicalGrid.

CosmoCentral.ComputeWeightFunction — Method

ComputeWeightFunction(z::Float64, i::Int64, ConvolvedDensity::AbstractConvolvedDensity,
AnalitycalDensity::AnalitycalDensity, InstrumentResponse::InstrumentResponse,
w0waCDMCosmology::w0waCDMCosmology, GCWeightFunction::GCWeightFunction)

This function returns the Galaxy Clustering Weight function for a given redshift $z$ and tomographic bin $i$.

CosmoCentral.ComputeWeightFunctionGrid! — Method

ComputeWeightFunctionGrid!(GCWeightFunction::GCWeightFunction,
ConvolvedDensity::AbstractConvolvedDensity, CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities, Cosmology::AbstractCosmology)

This function evaluates the Galaxy Clustering Weight Function over the $z$ grid and for all tomographic bins $i$.

For instance, here we plot the Galaxy Clustering weight function with a piecewise bias

PiecewiseBias = CosmoCentral.PiecewiseBias()
GCWeightFunction = CosmoCentral.GCWeightFunction(WeightFunctionArray =
zeros(length(ConvolvedDensity.DensityNormalizationArray), length(CosmologicalGrid.ZArray)))

CosmoCentral.ComputeBiasGrid!(CosmologicalGrid, GCWeightFunction, ConvolvedDensity)
CosmoCentral.ComputeWeightFunctionGrid!(GCWeightFunction, ConvolvedDensity,
CosmologicalGrid, BackgroundQuantities, w0waCDMCosmology)
p = Plots.plot(xlabel=L"z", ylabel=L"W_i^g(z)\,\left[\mathrm{Mpc}^{-1}\right]")
for i in 1:10
Plots.plot!(p, CosmologicalGrid.ZArray, GCWeightFunction.WeightFunctionArray[i,:],
    labels=(L"i=%$i"),  linewidth=3)
end
p

Computing $b(z)$ and $W_i^g(z)$ is quite fast

@benchmark CosmoCentral.ComputeBiasGrid!(CosmologicalGrid, GCWeightFunction, ConvolvedDensity)

BenchmarkTools.Trial: 7523 samples with 1 evaluation.
 Range (min … max):  476.000 μs …  25.209 ms  ┊ GC (min … max): 0.00% … 96.67%
 Time  (median):     600.600 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   659.423 μs ± 837.774 μs  ┊ GC (mean ± σ):  4.56% ±  3.53%

       ▅▂█▅█▅▇▄▆▂▃▁▃▁                                            
  ▂▂▅▇▇██████████████▇█▆▇▇▆▆▅▅▅▅▄▄▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▁▂▂▁▁▁▂▁▁▁▁▁▁▁ ▄
  476 μs           Histogram: frequency by time         1.01 ms <

 Memory estimate: 179.70 KiB, allocs estimate: 8501.

@benchmark CosmoCentral.ComputeWeightFunctionGrid!(GCWeightFunction, ConvolvedDensity,
CosmologicalGrid, BackgroundQuantities, w0waCDMCosmology)

BenchmarkTools.Trial: 4313 samples with 1 evaluation.
 Range (min … max):  723.202 μs … 28.429 ms  ┊ GC (min … max): 0.00% … 96.39%
 Time  (median):       1.024 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):     1.152 ms ±  1.625 ms  ┊ GC (mean ± σ):  8.94% ±  6.14%

    █▇█▆▅▃▃▂▁▂▁▃▅▄▃▄▅▅▃▁▂▃▁▁                                    
  ▅▄████████████████████████▇█▆▆▆▅▄▄▄▃▂▂▃▂▂▂▂▂▂▂▁▁▂▁▂▁▁▁▁▁▁▁▁▁ ▄
  723 μs          Histogram: frequency by time         1.87 ms <

 Memory estimate: 547.53 KiB, allocs estimate: 30021.

Weak Lensing

The expression of the Weak Lensing Weight Function is given by:

\[W_{i}^{\gamma}(z)=\frac{3}{2}\left(\frac{H_{0}}{c}\right)^{2} \Omega_{\mathrm{m}, 0}(1+z) r(z) \widetilde{W}_{i}(z)\]

where $\widetilde{W}_{i}(z)$ is the Lensing Efficiency, whose expression is given by

\[\widetilde{W}_{i}(z)=\int_{z}^{z_{\max }} \mathrm{d} z^{\prime} n_{i}\left(z^{\prime}\right)\left[1-\frac{\tilde{r}(z)}{\tilde{r}\left(z^{\prime}\right)}\right]\]

CosmoCentral.WLWeightFunction — Type

WLWeightFunction()

This struct contains the array with the Lensing Efficiency and Weak Lensing Weight Function values for all tomographic bins and redshift values in the CosmologicalGrid

CosmoCentral.ComputeWeightFunction — Method

ComputeWeightFunction(z::Float64, i::Int64,ConvolvedDensity::AbstractConvolvedDensity,
AnalitycalDensity::AnalitycalDensity, InstrumentResponse::InstrumentResponse,
Cosmology::AbstractCosmology, CosmologicalGrid::CosmologicalGrid,
LensingFunction::WLWeightFunction)

This function returns the Weak Lensing Weight Function, for a given redshift $z$ and tomographic bin $i$.

CosmoCentral.ComputeWeightFunctionGrid! — Method

ComputeWeightFunctionGrid!(LensingFunction::WLWeightFunction,
ConvolvedDensity::AbstractConvolvedDensity, CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities, Cosmology::AbstractCosmology)

This function evaluates the Weak Lensing Weight Function over the $z$ grid and for all tomographic bins $i$.

CosmoCentral.ComputeLensingEfficiency — Method

ComputeLensingEfficiency(z::Float64, i::Int64,
ConvolvedDensity::AbstractConvolvedDensity, AnalitycalDensity::AnalitycalDensity,
InstrumentResponse::InstrumentResponse, w0waCDMCosmology::w0waCDMCosmology,
CosmologicalGrid::CosmologicalGrid, ::WLWeightFunction)

This function returns the Lensing efficiency, for a given redshift $z$ and tomographic bin $i$.

CosmoCentral.ComputeLensingEfficiencyGrid! — Method

ComputeLensingEfficiencyGrid!(LensingFunction::WLWeightFunction,
AnalitycalDensity::AnalitycalDensity, InstrumentResponse::InstrumentResponse,
ConvolvedDensity::AbstractConvolvedDensity, CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities, Cosmology::AbstractCosmology,
::StandardLensingEfficiency)

This function evaluates the Lensing Efficiency over the $z$ grid and for all tomographic bins $i$.

CosmoCentral.ComputeLensingEfficiencyGrid! — Method

ComputeLensingEfficiencyGrid!(
LensingFunction::WLWeightFunction, ConvolvedDensity::AbstractConvolvedDensity,
CosmologicalGrid::CosmologicalGrid, BackgroundQuantities::BackgroundQuantities,
w0waCDMCosmology::w0waCDMCosmology, ::CustomLensingEfficiency)

This function evaluates the Lensing Efficiency over the $z$ grid and for all tomographic bins $i$.

Here we plot the Weak Lensing weight function. In particular, the solid lines are pure shear, while the dashed lines includes the Intrinsic Alignment contribution.

WLWeightFunction = CosmoCentral.WLWeightFunction(WeightFunctionArray = zeros(length(ConvolvedDensity.DensityNormalizationArray), length(CosmologicalGrid.ZArray)), LensingEfficiencyArray = zeros(length(ConvolvedDensity.DensityNormalizationArray), length(CosmologicalGrid.ZArray)))
CosmoCentral.ComputeLensingEfficiencyGrid!(
    WLWeightFunction, ConvolvedDensity,
    CosmologicalGrid,
    BackgroundQuantities,
    w0waCDMCosmology, CosmoCentral.CustomLensingEfficiency())
CosmoCentral.ComputeWeightFunctionGrid!(WLWeightFunction, ConvolvedDensity, CosmologicalGrid, BackgroundQuantities, w0waCDMCosmology)
p = Plots.plot(xlabel=L"z", ylabel=L"W_i^\gamma(z)\,\left[\mathrm{Mpc}^{-1}\right]")
for i in 1:10
Plots.plot!(p, CosmologicalGrid.ZArray, WLWeightFunction.WeightFunctionArray[i,:],
    labels=(L"i=%$i"),  linewidth=3)
end

CosmoCentral.ComputeIntrinsicAlignmentGrid!(CosmologicalGrid, WLWeightFunction, ConvolvedDensity, BackgroundQuantities, w0waCDMCosmology)
CosmoCentral.ComputeWeightFunctionGrid!(WLWeightFunction, ConvolvedDensity, CosmologicalGrid, BackgroundQuantities, w0waCDMCosmology)
for i in 1:10
Plots.plot!(p, CosmologicalGrid.ZArray, WLWeightFunction.WeightFunctionArray[i,:],
linewidth=3, linestyle = :dash)
end
p

Computations related to Weak Lensing are a bit slower, due to the nested integrals inside $W_i^\gamma(z)$

@benchmark CosmoCentral.ComputeLensingEfficiencyGrid!(WLWeightFunction, ConvolvedDensity,
CosmologicalGrid, BackgroundQuantities, w0waCDMCosmology,
CosmoCentral.CustomLensingEfficiency())

BenchmarkTools.Trial: 986 samples with 1 evaluation.
 Range (min … max):  4.113 ms … 14.512 ms  ┊ GC (min … max): 0.00% … 65.72%
 Time  (median):     4.806 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.058 ms ±  1.233 ms  ┊ GC (mean ± σ):  3.42% ±  8.98%

   ▂▅██▇▆▄▂▁▁▁                                                
  ▇███████████▇▆▇▆▁▅▄▅▄▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▄▅▅▁▄▅▆ █
  4.11 ms      Histogram: log(frequency) by time     12.7 ms <

 Memory estimate: 2.93 MiB, allocs estimate: 522.