Weight Function

Galaxy Clustering

The expression of the Galaxy Clustering Weight Function is given by:

\[W_{i}^{\mathrm{G}}(z)=b_{i}(z) n_{i}(z) \frac{H(z)}{c}\]

CosmoCentral.GCWeightFunction — Type

GCWeightFunction()

This struct contains the array with the Galaxy Bias and Galaxy Clustering Weight Function values for all tomographic bins and redshift values in the CosmologicalGrid.

CosmoCentral.ComputeWeightFunction — Method

ComputeWeightFunction(z::Float64, i::Int64, ConvolvedDensity::AbstractConvolvedDensity,
AnalitycalDensity::AnalitycalDensity, InstrumentResponse::InstrumentResponse,
w0waCDMCosmology::w0waCDMCosmology, GCWeightFunction::GCWeightFunction)

This function returns the Galaxy Clustering Weight function for a given redshift $z$ and tomographic bin $i$.

CosmoCentral.ComputeWeightFunctionGrid! — Method

ComputeWeightFunctionGrid!(GCWeightFunction::GCWeightFunction,
ConvolvedDensity::AbstractConvolvedDensity, CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities, w0waCDMCosmology::w0waCDMCosmology)

This function evaluates the Galaxy Clustering Weight Function over the $z$ grid and for all tomographic bins $i$.

Weak Lensing

The expression of the Weak Lensing Weight Function is given by:

\[W_{i}^{\gamma}(z)=\frac{3}{2}\left(\frac{H_{0}}{c}\right)^{2} \Omega_{\mathrm{m}, 0}(1+z) r(z) \widetilde{W}_{i}(z)\]

where $\widetilde{W}_{i}(z)$ is the Lensing Efficiency, whose expression is given by

\[\widetilde{W}_{i}(z)=\int_{z}^{z_{\max }} \mathrm{d} z^{\prime} n_{i}\left(z^{\prime}\right)\left[1-\frac{\tilde{r}(z)}{\tilde{r}\left(z^{\prime}\right)}\right]\]

CosmoCentral.WLWeightFunction — Type

WLWeightFunction()

This struct contains the array with the Lensing Efficiency and Weak Lensing Weight Function values for all tomographic bins and redshift values in the CosmologicalGrid

CosmoCentral.ComputeWeightFunction — Method

ComputeWeightFunction(z::Float64, i::Int64,ConvolvedDensity::AbstractConvolvedDensity,
AnalitycalDensity::AnalitycalDensity, InstrumentResponse::InstrumentResponse,
w0waCDMCosmology::w0waCDMCosmology, CosmologicalGrid::CosmologicalGrid,
LensingFunction::WLWeightFunction)

This function returns the Weak Lensing Weight Function, for a given redshift $z$ and tomographic bin $i$.

CosmoCentral.ComputeWeightFunctionGrid! — Method

ComputeWeightFunctionGrid!(LensingFunction::WLWeightFunction,
ConvolvedDensity::AbstractConvolvedDensity, CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities, w0waCDMCosmology::w0waCDMCosmology)

This function evaluates the Weak Lensing Weight Function over the $z$ grid and for all tomographic bins $i$.

CosmoCentral.ComputeLensingEfficiency — Method

ComputeLensingEfficiency(z::Float64, i::Int64,
ConvolvedDensity::AbstractConvolvedDensity, AnalitycalDensity::AnalitycalDensity,
InstrumentResponse::InstrumentResponse, w0waCDMCosmology::w0waCDMCosmology,
CosmologicalGrid::CosmologicalGrid, ::WLWeightFunction)

This function returns the Lensing efficiency, for a given redshift $z$ and tomographic bin $i$.

CosmoCentral.ComputeLensingEfficiencyGrid! — Method

ComputeLensingEfficiencyGrid!(
    LensingFunction::WLWeightFunction,
    AnalitycalDensity::AnalitycalDensity,
    InstrumentResponse::InstrumentResponse,
    ConvolvedDensity::AbstractConvolvedDensity,
    CosmologicalGrid::CosmologicalGrid,
    BackgroundQuantities::BackgroundQuantities,
    w0waCDMCosmology::w0waCDMCosmology,
    ::StandardLensingEfficiency)

This function evaluates the Lensing Efficiency over the $z$ grid and for all tomographic bins $i$.

CosmoCentral.ComputeLensingEfficiencyGrid! — Method

ComputeLensingEfficiencyGrid!(
LensingFunction::WLWeightFunction, ConvolvedDensity::AbstractConvolvedDensity,
CosmologicalGrid::CosmologicalGrid, BackgroundQuantities::BackgroundQuantities,
w0waCDMCosmology::w0waCDMCosmology, ::CustomLensingEfficiency)

This function evaluates the Lensing Efficiency over the $z$ grid and for all tomographic bins $i$.