Weight Function

Galaxy Clustering

CosmoCentral.GCWeightFunctionStruct — Type

GCWeightFunctionStruct()

This struct contains the array with the Galaxy Clustering Weight function values for all tomographic bins and redshift values in the CosmologicalGridStruct.

CosmoCentral.ComputeWeightFunction — Function

ComputeWeightFunction(z::Float64, i::Int64,
    ConvolvedDensity::ConvolvedDensity,
    w0waCDMCosmology::w0waCDMCosmology)

This function returns the Galaxy Clustering Weight function for a given redshift $z$ and tomographic bin $i$.

CosmoCentral.ComputeWeightFunctionOverGrid — Function

ComputeWeightFunctionOverGrid(
GCWeightFunction::GCWeightFunction, AnalitycalDensity::AnalitycalDensity,
InstrumentResponse::InstrumentResponse, ConvolvedDensity::ConvolvedDensity,
Bias::Bias, CosmologicalGrid::CosmologicalGrid,
BackgroundQuantities::BackgroundQuantities,
w0waCDMCosmology::w0waCDMCosmology)

This function evaluates the Galaxy Clustering Weight Function over the $z$ grid and for all tomographic bins $i$.