The CosmoCentral Module

Missing docstring.

Missing docstring for CosmoCentral. Check Documenter's build log for details.

Module Index

CosmoCentral.ComputeAdimensionalHubbleFactor

CosmoCentral.ComputeComovingDistance

CosmoCentral.ComputeHubbleFactor

Detailed API

CosmoCentral.ComputeAdimensionalHubbleFactor — Function

ComputeAdimensionalHubbleFactor(z, $w0$, wa, ΩM, ΩDE, Ωk, Ωr)

This function, given the value of the cosmological parameters, evaluate the Adimensional Hubble Factor for w0waCDM cosmologies

...

Arguments

n::Integer: the number of elements to compute.

dim::Integer=1: the dimensions along which to perform the computation.

...

Notes

Notes can go here

Examples

```julia julia> five = plusTwo(3) 5

CosmoCentral.ComputeComovingDistance — Function

ComputeComovingDistance(z, H0, w0, wa, ΩM, ΩDE, Ωk, Ωr)

This function, given the value of the cosmological parameters, evaluate the Comoving Distance for w0waCDM cosmologies

CosmoCentral.ComputeHubbleFactor — Function

ComputeHubbleFactor(z, H0, w0, wa, ΩM, ΩDE, Ωk, Ωr)

This function, given the value of the cosmological parameters, evaluate the Hubble Factor for w0waCDM cosmologies