The CosmoCentral Module

Module Index

CosmoCentral.ComputeAdimensionalHubbleFactor

CosmoCentral.ComputeComovingDistance

CosmoCentral.ComputeHubbleFactor

Detailed API

CosmoCentral.ComputeAdimensionalHubbleFactor — Method

ComputeAdimensionalHubbleFactor($z$, params)

This function, given the value of the cosmological parameters, evaluate the Adimensional Hubble Factor for $w_0 w_a$CDM cosmologies.

...

Arguments

z::Float64 the redshift value at which evaluate the Adimensional Hubble Factor

params::w0waCDMCosmology a collection of cosmological parameters, whose expression is given by

\[E(z)=\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Notes

This expression is valid only for the CPL parameterization

CosmoCentral.ComputeComovingDistance — Method

ComputeComovingDistance($z$, params)

This function, given the value of the cosmological parameters, evaluate the Comoving Distance. It is evaluated as

\[r(z)=\frac{c}{H_0}\int_0^z \frac{dx}{E(x)}\]

CosmoCentral.ComputeHubbleFactor — Method

ComputeHubbleFactor($z$, params)

...

Arguments

z::Float64 the redshift value at which evaluate the Adimensional Hubble Factor

params::w0waCDMCosmology a collection of cosmological parameters

This function, given the value of the cosmological parameters, evaluate the Hubble Factor for $w_0 w_a$CDM cosmologies, whose expression is given by

\[H(z)=H_0\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Notes

This expression is valid only for the CPL parameterization