The CosmoCentral Module

Module Index

CosmoCentral.AnalitycalDensityStruct

CosmoCentral.ConvolvedDensityStruct

CosmoCentral.CosmoGridStruct

CosmoCentral.InstrumentResponseStruct

CosmoCentral.w0waCDMStruct

CosmoCentral.ComputeAdimensionalHubbleFactor

CosmoCentral.ComputeComovingDistance

CosmoCentral.ComputeConvolvedDensityFunction

CosmoCentral.ComputeDensityFunction

CosmoCentral.ComputeDensityFunctionConvolvedGrid

CosmoCentral.ComputeHubbleFactor

CosmoCentral.ComputeInstrumentResponse

CosmoCentral.LogSpaced

CosmoCentral.NormalizeAnalitycalDensityStruct

CosmoCentral.NormalizeConvolvedDensityStruct

Detailed API

CosmoCentral.AnalitycalDensityStruct — Type

AnalitycalDensityStruct(z0::Float64 = 0.9/sqrt(2.), zmin::Float64 = 0.001,
zmax::Float64 = 2.5, surfacedensity::Float64 = 30.,
normalization::Float64 = 1.)

This struct contains the parameters of the source galaxy density as given by the official Euclid forecast, whose expression is given by:

\[n(z)\propto\left(\frac{z}{z_0}\right)^2 \exp{\left(-\left(\frac{z}{z_0}\right)^{-3/2}\right)}\]

The parameters contained in this struct are

$z_{min}$ and $z_{max}$, the minimum and the maximum redshift considered

$z_0$, the parameter present in the galaxy distribution

surfacedensity , the value of the galaxy source density integrated between $z_{min}$ and $z_{max}$

normalization, the value of parameter which multiplies the source dennsity in order to match the correct surface density

CosmoCentral.ConvolvedDensityStruct — Type

ConvolvedDensityStruct(AnalitycalDensity::AnalitycalDensity = AnalitycalDensityStruct(),
InstrumentResponse::InstrumentResponse = InstrumentResponseStruct()
zbinarray::Vector{Float64} = Array([0.001, 0.418, 0.560, 0.678, 0.789, 0.900, 1.019, 1.155, 1.324, 1.576, 2.50])
densityarraynormalization::Vector{Float64} = ones(length(zbinarray)-1)
densitygridarray::AbstractArray{Float64, 2} = ones(length(zbinarray)-1, 300))

This function evaluate the normalization of the convolved density function.

CosmoCentral.CosmoGridStruct — Type

w0waCDMParameters

This struct contains the value of the cosmological parameters for $w_0 w_a$CDM cosmologies:

$w_0$ and $w_a$, the parameters in the CPL parameterization

$\Omega_M$, $\Omega_B$, $\Omega_{DE}$, $\Omega_R$, and $\Omega_k$ the density parameters for matter, baryons, Dark Energy, radiation, curvature

$n_s$, the scalar spectral index

$M_\nu$, the sum of the neutrino mass eigenstates in eV

$\sigma_8$, the amplitude of the scalar fluctuations

$H_0$, the value of the Hubble paramater

CosmoCentral.InstrumentResponseStruct — Type

InstrumentResponseStruct(cb::Float64 = 1.0, zb::Float64 = 0.0,
σb::Float64 = 0.05, co::Float64 = 1.0, zo::Float64 = 0.1,
σo::Float64 = 0.05, fout::Float64 = 0.1)

When we measure the redshift of a galaxy with redshit $z$, we will measure a redshift $z_p$ with a probability given by the following expression:

\[p(z_p|z) = \frac{1-f_{out}}{\sqrt{2 \pi} \sigma_{b}(1+z)} \exp \left( -\frac{1}{2}\left(\frac{z-c_{b} z_{b}-z_{b}}{\sigma_{b}(1+z)}\right)^{2} \right) + \frac{f_{out}}{\sqrt{2 \pi} \sigma_{\mathrm{o}}(1+z)} \exp \left(-\frac{1}{2}\left(\frac{z-c_{o} z_{p}-z_{o}}{\sigma_{o}(1+z)} \right)^{2}\right)\]

This struct contains all these parameters.

CosmoCentral.w0waCDMStruct — Type

w0waCDMParameters(w0::Float64 = -1, wa::Float64 = 0, ΩM::Float64 = 0.32,
ΩB::Float64  = 0.05, ΩDE::Float64 = 0.68, Ωk::Float64  = 0.,
Ωr::Float64  = 0., ns::Float64  = 0.96, Mν::Float64  = 0.06,
σ8::Float64  = 0.816, H0::Float64  = 67.)

This struct contains the value of the cosmological parameters for $w_0 w_a$CDM cosmologies:

$w_0$ and $w_a$, the parameters in the CPL parameterization

$\Omega_M$, $\Omega_B$, $\Omega_{DE}$, $\Omega_R$, and $\Omega_k$ the density parameters for matter, baryons, Dark Energy, radiation, curvature

$n_s$, the scalar spectral index

$M_\nu$, the sum of the neutrino mass eigenstates in eV

$\sigma_8$, the amplitude of the scalar fluctuations

$H_0$, the value of the Hubble paramater

CosmoCentral.ComputeAdimensionalHubbleFactor — Method

ComputeAdimensionalHubbleFactor(z::Float64, params::w0waCDMCosmology)

This function, given the value of the cosmological parameters, evaluate the Adimensional Hubble Factor for $w_0 w_a$CDM cosmologies. The analitycal expression is given by:

\[E(z)=\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Arguments

z::Float64 the redshift value

params::w0waCDMCosmology, a collection of cosmological parameters

Example

julia> params = CosmoCentral.w0waCDMStruct()
julia> CosmoCentral.ComputeAdimensionalHubbleFactor(0., params)
1.0

Warning

This expression is valid only for the CPL parameterization of the Dark Energy Equation of State.

CosmoCentral.ComputeComovingDistance — Method

ComputeComovingDistance(z::Float64, params::w0waCDMCosmology)

This function, given the value of the cosmological parameters, evaluate the Comoving Distance. It is evaluated as:

\[r(z)=\frac{c}{H_0}\int_0^z \frac{dx}{E(x)}\]

Example

julia> params = CosmoCentral.w0waCDMStruct(H0=67.)
julia> CosmoCentral.ComputeAdimensionalHubbleFactor(0., params)
67.0

CosmoCentral.ComputeConvolvedDensityFunction — Method

ComputeDensityFunction(z::Float64, i::Int64, convolveddensity::ConvolvedDensity)

This function computes the Convolved density function for a single bin at a given redshift $z$.

CosmoCentral.ComputeDensityFunction — Method

ComputeDensityFunction(z::Float64, AnalitycalDensity::AnalitycalDensity = AnalitycalDensityStruct())

This function returns the source density for a given redshift $z$.

CosmoCentral.ComputeDensityFunctionConvolvedGrid — Method

ComputeDensityFunction(CosmoGrid::CosmoGrid, ConvolvedDensityStruct::ConvolvedDensity)

This function computes the convolved density function for all tomographic bins on the $z$-grid provided by CosmoGrid.

CosmoCentral.ComputeHubbleFactor — Method

ComputeHubbleFactor(z::Float64, params::w0waCDMCosmology)

Arguments

z::Float64, the redshift value

params::w0waCDMCosmology, a collection of cosmological parameters

This function, given the value of the cosmological parameters, evaluate the Hubble Factor for $w_0 w_a$CDM cosmologies, whose expression is given by

\[H(z)=H_0\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+ \Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Example

julia> params = CosmoCentral.w0waCDMStruct(H0=67.)
julia> CosmoCentral.ComputeAdimensionalHubbleFactor(0., params)
67.0

CosmoCentral.ComputeInstrumentResponse — Method

ComputeInstrumentResponse(z::Float64, zp::Float64, instrumentresponse::InstrumentResponse)

This function computes the probability that we actually measure a redshift $z_p$ if the real redshift is $z$.

CosmoCentral.LogSpaced — Method

LogSpaced(minarray::Float64, maxarray::Float64, n::Int64)

This function evaluates $n$ points, logarithmically spaced between $z_{minarray}$ and $z_{maxarray}$.

CosmoCentral.NormalizeAnalitycalDensityStruct — Method

NormalizeAnalitycalDensityStruct(densityparameters::AnalitycalDensity)

This function normalize AnalitycalDensityStruct in order to have the correct value of the surface density once integrated.

CosmoCentral.NormalizeConvolvedDensityStruct — Method

NormalizeConvolvedDensityStruct(convolveddensity::ConvolvedDensity)

This function normalizes ConvolvedDensity such that the integrals of the convolved densities are normalized to 1.