The CosmoCentral Module

Module Index

CosmoCentral.AnalitycalDensityStruct

CosmoCentral.ConvolvedDensityStruct

CosmoCentral.CosmoGridStruct

CosmoCentral.w0waCDMStruct

CosmoCentral.ComputeAdimensionalHubbleFactor

CosmoCentral.ComputeComovingDistance

CosmoCentral.ComputeConvolvedDensityFunction

CosmoCentral.ComputeDensityFunction

CosmoCentral.ComputeDensityFunctionConvolvedGrid

CosmoCentral.ComputeHubbleFactor

CosmoCentral.ComputeInstrumentResponse

CosmoCentral.NormalizeAnalitycalDensityStruct

CosmoCentral.NormalizeConvolvedDensityStruct

Detailed API

CosmoCentral.AnalitycalDensityStruct — Type

AnalitycalDensityStruct

This struct contains the parameters of the source galaxy density as given by the official Euclid forecast, whose expression is given by:

\[n(z)\propto(\frac{z}{z_0})^2 \exp{-(\frac{z}{z_0})^{-3/2}}\]

The parameters contained in this struct are

$z_{min}$ and $z_{max}$, the minimum and the maximum redshift considered

$z_0$, the parameter present in the galaxy distribution

surfacedensity, the value of the galaxy source density integrated between $z_{min}$ and $z_{max}$

normalization, the value of parameter which multiplies the source dennsity in order to match the correct surface density

CosmoCentral.ConvolvedDensityStruct — Type

\[p(z_p|z) = \frac{1-f_{\mathrm{out}}}{\sqrt{2 \pi} \sigma_{\mathrm{b}}(1+z)} \exp \left\{-\frac{1}{2}\left[\frac{z-c_{\mathrm{b}} z_{\mathrm{p}}-z_{\mathrm{b}}}{\sigma_{\mathrm{b}}(1+z)}\right]^{2}\right\} \, + \, \frac{f_{\mathrm{out}}}{\sqrt{2 \pi} \sigma_{\mathrm{o}}(1+z)} \exp \left\{- rac{1}{2}\left[\frac{z-c_{\mathrm{o}} z_{\mathrm{p}}-z_{\mathrm{o}}}{\sigma_{\mathrm{o}}(1+z)}\right]^{2}\right\}\]

CosmoCentral.CosmoGridStruct — Type

w0waCDMParameters

This struct contains the value of the cosmological parameters for $w_0 w_a$CDM cosmologies:

$w_0$ and $w_a$, the parameters in the CPL parameterization

$\Omega_M$, $\Omega_B$, $\Omega_{DE}$, $\Omega_R$, and $\Omega_k$ the density parameters for matter, baryons, Dark Energy, radiation, curvature

$n_s$, the scalar spectral index

$M_\nu$, the sum of the neutrino mass eigenstates in eV

$\sigma_8$, the amplitude of the scalar fluctuations

$H_0$, the value of the Hubble paramater

CosmoCentral.w0waCDMStruct — Type

w0waCDMParameters

This struct contains the value of the cosmological parameters for $w_0 w_a$CDM cosmologies:

$w_0$ and $w_a$, the parameters in the CPL parameterization

$\Omega_M$, $\Omega_B$, $\Omega_{DE}$, $\Omega_R$, and $\Omega_k$ the density parameters for matter, baryons, Dark Energy, radiation, curvature

$n_s$, the scalar spectral index

$M_\nu$, the sum of the neutrino mass eigenstates in eV

$\sigma_8$, the amplitude of the scalar fluctuations

$H_0$, the value of the Hubble paramater

CosmoCentral.ComputeAdimensionalHubbleFactor — Method

ComputeAdimensionalHubbleFactor($z$, params)

This function, given the value of the cosmological parameters, evaluate the Adimensional Hubble Factor for $w_0 w_a$CDM cosmologies.

...

Arguments

z::Float64 the redshift value at which evaluate the Adimensional Hubble Factor

params::w0waCDMCosmology a collection of cosmological parameters, whose expression is given by

\[E(z)=\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Notes

This expression is valid only for the CPL parameterization

CosmoCentral.ComputeComovingDistance — Method

ComputeComovingDistance($z$, params)

This function, given the value of the cosmological parameters, evaluate the Comoving Distance. It is evaluated as

\[r(z)=\frac{c}{H_0}\int_0^z \frac{dx}{E(x)}\]

CosmoCentral.ComputeConvolvedDensityFunction — Method

ComputeDensityFunction(CosmoGrid, ConvolvedDensity)

This function computes the Convolved density function for a single bin at a given $z$ redshift.

CosmoCentral.ComputeDensityFunction — Method

ComputeDensityFunction($z$, AnalitycalDensity)

This function returns the source density for a given redshift $z$

CosmoCentral.ComputeDensityFunctionConvolvedGrid — Method

ComputeDensityFunction(CosmoGrid, ConvolvedDensity)

This function computes the convolved density function for all tomographic bins on the $z$-grid provided by CosmoGrid.

CosmoCentral.ComputeHubbleFactor — Method

ComputeHubbleFactor($z$, params)

...

Arguments

z::Float64 the redshift value at which evaluate the Adimensional Hubble Factor

params::w0waCDMCosmology a collection of cosmological parameters

This function, given the value of the cosmological parameters, evaluate the Hubble Factor for $w_0 w_a$CDM cosmologies, whose expression is given by

\[H(z)=H_0\sqrt{\Omega_M(1+z)^3+\Omega_R(1+z)^4+\Omega_{DE}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Notes

This expression is valid only for the CPL parameterization

CosmoCentral.ComputeInstrumentResponse — Method

ComputeInstrumentResponse($z$, `z_p, InstrumentResponseStruct)

This function returns the instrument response.

CosmoCentral.NormalizeAnalitycalDensityStruct — Method

NormalizeAnalitycalDensityStruct(AnalitycalDensity)

This function normalize AnalitycalDensityStruct in order to have the correct value of the surface density once integrated.

CosmoCentral.NormalizeConvolvedDensityStruct — Method

ComputeDensityFunction(ConvolvedDensity)

This function normalizes ConvolvedDensity such that the integrals of the convolved densities are normalized to 1.